make return type of ArtifactList::count an int again
Linux gcc won't implicitely convert size_t to uint when streaming.

+++ src/lib/buildgraph/artifactlist.h
@@ -91,9 +91,9 @@ public:
         return m_data.empty();
     }
 
-    size_t count() const
+    int count() const
     {
-        return m_data.size();
+        return (int)m_data.size();
     }
 
     void reserve(int)
